Move Custom*Scalar interfaces into the API module
