+++ server/api/src/main/java/io/smallrye/graphql/api/CustomStringScalar.java
@@ -1,4 +1,4 @@
-package io.smallrye.graphql.scalar.custom;
+package io.smallrye.graphql.api;
 
 /**
  * A base class for all CustomScalars that are based on GraphQL's String.
